simplify langchain component implementaiton
