+++ sdks/python/src/stepflow_py/http_server.py
@@ -359,3 +359,7 @@ class StepflowHttpServer:
         )
         server = uvicorn.Server(config)
         await server.serve()
+
+    def langchain_component(self, *args, **kwargs):
+        """Delegate langchain_component registration to the underlying server."""
+        return self.server.langchain_component(*args, **kwargs)
