+++ sdks/python/src/stepflow_py/server.py
@@ -53,6 +53,14 @@ from stepflow_py.generated_protocol import (
 )
 from stepflow_py.udf import udf
 
+# Check if LangChain is available
+try:
+    from stepflow_py.langchain_components import register_langchain_components
+
+    _HAS_LANGCHAIN = True
+except ImportError:
+    _HAS_LANGCHAIN = False
+
 
 @dataclass
 class ComponentEntry:
@@ -91,8 +99,9 @@ class StepflowServer:
         self._components: dict[str, ComponentEntry] = {}
         self._initialized = False
 
-        # Add LangChain registry functionality
-        self._add_langchain_support()
+        # Add LangChain registry functionality if available
+        if _HAS_LANGCHAIN:
+            register_langchain_components(self)
 
         if include_builtins:
             # Register the UDF component
@@ -376,18 +385,166 @@ class StepflowServer:
             traceback.print_exc(file=sys.stderr)
             raise StepflowExecutionError(f"Component execution failed: {str(e)}") from e
 
-    def _add_langchain_support(self):
-        """Add LangChain integration support to the server."""
-        try:
-            from stepflow_py.langchain_components import register_langchain_components
-            from stepflow_py.langchain_registry import add_langchain_component_method
+    def langchain_component(
+        self,
+        func: Callable | None = None,
+        *,
+        name: str | None = None,
+        description: str | None = None,
+        execution_mode: str = "invoke",
+    ):
+        """
+        Decorator to register a LangChain runnable factory as a StepFlow component.
 
-            # Add the langchain_component decorator method
-            add_langchain_component_method(self.__class__)
+        The decorated function should return a LangChain Runnable instance.
+        The resulting component will execute the runnable with the provided input.
 
-            # Register built-in LangChain components
-            register_langchain_components(self)
+        Args:
+            func: The function to register (provided by the decorator)
+            name: Optional name for the component. If not provided, uses the
+                function name
+            description: Optional description. If not provided, uses the
+                function's docstring
+            execution_mode: Default execution mode ("invoke", "batch", "stream")
 
-        except ImportError:
-            # LangChain not available, skip adding support
-            pass
+        Example:
+            @server.langchain_component(name="my_chain")
+            def create_my_chain() -> Runnable:
+                return prompt | llm | parser
+
+        Raises:
+            StepflowExecutionError: If LangChain is not available
+        """
+        if not _HAS_LANGCHAIN:
+            raise StepflowExecutionError(
+                "LangChain integration requires langchain-core. "
+                "Install with: pip install stepflow-py[langchain]"
+            )
+
+        def decorator(f: Callable) -> Callable:
+            from stepflow_py.langchain_integration import (
+                check_langchain_available,
+                create_runnable_config,
+                get_runnable_schemas,
+            )
+
+            check_langchain_available()
+
+            component_name = name or f.__name__
+            if not component_name.startswith("/"):
+                component_name = f"/{component_name}"
+
+            # Validate execution mode
+            if execution_mode not in ["invoke", "batch", "stream"]:
+                raise ValueError(
+                    f"Invalid execution_mode '{execution_mode}'. "
+                    "Must be 'invoke', 'batch', or 'stream'"
+                )
+
+            # Create the runnable instance by calling the factory function
+            try:
+                runnable = f()
+            except Exception as e:
+                raise StepflowExecutionError(
+                    f"Failed to create LangChain runnable: {e}"
+                ) from e
+
+            # Validate that it's actually a runnable
+            try:
+                from langchain_core.runnables import Runnable
+
+                if not isinstance(runnable, Runnable):
+                    raise StepflowExecutionError(
+                        f"Function {f.__name__} must return a LangChain Runnable "
+                        f"instance, got {type(runnable)}"
+                    )
+            except ImportError:
+                raise StepflowExecutionError("LangChain not available") from None
+
+            # Get schemas from the runnable
+            try:
+                input_schema, output_schema = get_runnable_schemas(runnable)
+            except Exception:
+                # Fallback to generic schemas
+                input_schema = {"type": "object", "additionalProperties": True}
+                output_schema = {"type": "object", "additionalProperties": True}
+
+            # Define LangChainComponentInput class inline
+
+            class LangChainComponentInput(msgspec.Struct):
+                """Input structure for LangChain components."""
+
+                input: Any
+                config: dict[str, Any] | None = None
+                execution_mode: str = "invoke"
+
+            # Create the component function that will execute the runnable
+            async def langchain_component_executor(
+                component_input: LangChainComponentInput, context: StepflowContext
+            ) -> Any:
+                """Execute the registered LangChain runnable."""
+
+                # Validate execution mode (only invoke is supported)
+                exec_mode = component_input.execution_mode or execution_mode
+                if exec_mode != "invoke":
+                    raise StepflowExecutionError(
+                        f"Invalid execution_mode '{exec_mode}'. "
+                        "Only 'invoke' mode is supported."
+                    )
+
+                # Prepare input for LangChain runnable
+                langchain_input = component_input.input
+
+                # Create runnable config from StepFlow input
+                stepflow_input_dict = {
+                    "input": component_input.input,
+                    "config": component_input.config or {},
+                }
+                runnable_config = create_runnable_config(stepflow_input_dict, context)
+
+                # Execute the runnable
+                return await runnable.ainvoke(langchain_input, config=runnable_config)
+
+            # Set up the component input/output types and description
+            langchain_component_executor.__annotations__ = {
+                "component_input": LangChainComponentInput,
+                "context": StepflowContext,
+                "return": Any,
+            }
+
+            # Use provided description or function docstring
+            component_description = description or (
+                f.__doc__.strip() if f.__doc__ else None
+            )
+            if component_description:
+                langchain_component_executor.__doc__ = component_description
+
+            # Register the component using the server's component method
+            self.component(
+                langchain_component_executor,
+                name=component_name,
+                description=component_description,
+            )
+
+            # Store additional metadata about the LangChain component
+            if not hasattr(self, "_langchain_components"):
+                self._langchain_components: dict[str, dict[str, Any]] = {}
+
+            self._langchain_components[component_name] = {
+                "factory_function": f,
+                "runnable": runnable,
+                "input_schema": input_schema,
+                "output_schema": output_schema,
+                "execution_mode": execution_mode,
+                "description": component_description,
+            }
+
+            @wraps(f)
+            def wrapper(*args, **kwargs):
+                return f(*args, **kwargs)
+
+            return wrapper
+
+        if func is None:
+            return decorator
+        return decorator(func)
